feat: update tap to 18
BREAKING CHANGE: This update is not backwards compatible with previous versions of testing.

This also adds some tests to the skeleton itself showcases how to use mocking with built in plugins, and spies via the added sinon plugin. A .taprc file has also been added as an example of adding some tap configuration.